+++ backend/modules/locale/engine/model.php
@@ -1259,7 +1259,7 @@ class BackendLocaleModel
 	public static function insert(array $item)
 	{
 		// actions should be urlized
-		if($item['type'] == 'act' && urldecode($item['value']) != $item['value']) $item['value'] = SpoonFilter::urlise($item['value']);
+		if($item['type'] == 'act' && urldecode($item['value']) != $item['value']) $item['value'] = CommonUri::getUrl($item['value']);
 
 		// insert item
 		$item['id'] = (int) BackendModel::getContainer()->get('database')->insert('locale', $item);
@@ -1279,7 +1279,7 @@ class BackendLocaleModel
 	public static function update(array $item)
 	{
 		// actions should be urlized
-		if($item['type'] == 'act' && urldecode($item['value']) != $item['value']) $item['value'] = SpoonFilter::urlise($item['value']);
+		if($item['type'] == 'act' && urldecode($item['value']) != $item['value']) $item['value'] = CommonUri::getUrl($item['value']);
 
 		// update category
 		$updated = BackendModel::getContainer()->get('database')->update('locale', $item, 'id = ?', array($item['id']));