+++ backend/core/engine/model.php
@@ -886,8 +886,8 @@ class BackendModel extends BaseModel
 		// build URL
 		$URL = self::getURL($pageIdForURL, $language);
 
-		// set locale
-		FrontendLanguage::setLocale($language);
+		// set locale with force
+		FrontendLanguage::setLocale($language, true);
 
 		// append action
 		$URL .= '/' . urldecode(FL::act(SpoonFilter::toCamelCase($action)));