+++ backend/modules/tags/ajax/edit.php
@@ -27,6 +27,21 @@ class BackendTagsAjaxEdit extends BackendBaseAJAXAction
 
 		// validate id
 		if($id === 0) $this->output(self::BAD_REQUEST, null, 'no id provided');
+		if($tag === '') $this->output(self::BAD_REQUEST, null, BL::err('NameIsRequired'));
+
+		// check if tag exists
+		if(BackendTagsModel::existsTag($tag)) $this->output(self::BAD_REQUEST, null, BL::err('TagAlreadyExists'));
+
+		// build array
+		$item['id'] = $id;
+		$item['tag'] = SpoonFilter::htmlspecialchars($tag);
+		$item['url'] = BackendTagsModel::getURL(CommonUri::getUrl(SpoonFilter::htmlspecialcharsDecode($item['tag'])), $id);
+
+		// update
+		BackendTagsModel::update($item);
+
+		// output
+		$this->output(self::OK, $item, vsprintf(BL::msg('Edited'), array($item['tag'])));
 		
 		// validated id
 		else